refactor(analytics): separate reusable graph component
